renamed adapters package to ports
